clean up for public repository
