+++ mars_troughs/datapaths.py
@@ -3,7 +3,6 @@ Paths to data files and helper methods to load some of them.
 """
 from pathlib import Path
 from typing import List, Tuple
-
 import numpy as np
 import pandas as pd
