+++ mars_troughs/lag_model.py
@@ -1,12 +1,15 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
 """
-Models for the lag as a function of time.
+Custom lag models
 """
-from typing import Dict
-
 import numpy as np
-
-from mars_troughs.generic_model import ConstantModel, LinearModel
 from mars_troughs.model import Model
+from mars_troughs.generic_model import (ConstantModel, 
+                                        LinearModel,
+                                        QuadModel, 
+                                        CubicModel, 
+                                        PowerLawModel)
 
 
 class LagModel(Model):
@@ -30,8 +33,7 @@ class LagModel(Model):
             np.ndarray of the same size as time input containing values of lag.
         """
         return self.eval(time)
-
-
+    
 class ConstantLag(LagModel, ConstantModel):
     """
     The lag thickness is constant and does not depend on time.
@@ -52,10 +54,10 @@ class ConstantLag(LagModel, ConstantModel):
 class LinearLag(LagModel, LinearModel):
     """
     The lag thickness is linear in time. Lag changes as
-    lag(t) = constant + slope*t.
+    lag(t) = intercept + slope*t.
 
     Args:
-        constant (float, optional): default is 1 millimeter. The lag
+        intercept (float, optional): default is 1 millimeter. The lag
             thickness at time t=0 (present day).
         slope (float, optional): default is 1e-6 mm per year. The rate
             of change of the lag per time.
@@ -68,9 +70,55 @@ class LinearLag(LagModel, LinearModel):
     ):
         super().__init__()
         LinearModel.__init__(self, constant=constant, slope=slope)
-
-
-LAG_MODEL_MAP: Dict[str, Model] = {
-    "constant": ConstantLag,
-    "linear": LinearLag,
-}
+    
+class QuadraticLag(LagModel,QuadModel):
+    """
+    Lag is quadratic with time
+    """
+    
+    def __init__(
+        self,
+        constant: float = 1e-6,
+        slope: float = 1e-8,
+        quad: float = 1e-20,
+    ):
+        super().__init__()
+        QuadModel.__init__(self, constant=constant, 
+                                 slope=slope,
+                                 quad=quad)
+        
+class CubicLag(LagModel,CubicModel):
+    """
+    Lag is cubic with time
+    """
+    
+    def __init__(
+        self,
+        constant: float = 1e-6,
+        slope: float = 1e-8,
+        quad: float = 1e-20,
+        cubic: float = 1e-30,
+    ):
+        super().__init__()
+        CubicModel.__init__(self, constant=constant, 
+                                  slope=slope,
+                                  quad=quad,
+                                  cubic=cubic)
+    
+class PowerLawLag(LagModel,PowerLawModel):
+    """
+    Lag follows a power law function with time
+    """
+    def __init__(
+        self,
+        coeff: float = 5,
+        exponent: float = 1e-2,
+        ):
+        super().__init__()
+        PowerLawModel.__init__(self, coeff, exponent)
+        
+        
+    
+        
+    
+
