+++ scripts/cmSoftAgePrior.py
@@ -3,10 +3,21 @@ import argparse
 import pickle
 import numpy as np
 import mars_troughs as mt
-from mars_troughs.custom_lag_models import ConstantLag,LinearLag,QuadraticLag,CubicLag,PowerLawLag
-from mars_troughs.custom_acc_models import Linear_Insolation, Quadratic_Insolation,Cubic_Insolation,PowerLaw_Insolation
-from mars_troughs.custom_acc_models import Linear_Obliquity, Quadratic_Obliquity,Cubic_Obliquity,PowerLaw_Obliquity
-from mars_troughs.datapaths import load_insolation_data, load_obliquity_data
+from mars_troughs import (ConstantLag,
+                          LinearLag,
+                          QuadraticLag,
+                          CubicLag,
+                          PowerLawLag)
+from mars_troughs import (Linear_Insolation, 
+                          Quadratic_Insolation,
+                          Cubic_Insolation,
+                          PowerLaw_Insolation)
+from mars_troughs import (Linear_Obliquity, 
+                          Quadratic_Obliquity,
+                          Cubic_Obliquity,
+                          PowerLaw_Obliquity)
+from mars_troughs import (load_insolation_data, 
+                          load_obliquity_data)
 import sys
 
 def main():
@@ -62,8 +73,7 @@ def main():
     angle=5.0
     thin_by=args.thin_by
     mcmcobj=mt.softAgePriorMCMC(meanAge,maxSteps,thin_by,directory,tmp,
-                                 acc_model,lag_model, None, None, 
-                                 errorbar, angle)
+                                 acc_model,lag_model,errorbar, angle)
 
     
     filename=mcmcobj.filename+'_soft'+ str(int(args.meanAge))
